refactor(rule): collapse similarity switches into one helper
The processor had two near-identical switches over a SimilarityOutcome
discriminated union — one in evaluateAddressDataWithCoordinates and
one in evaluateTextualAddressWithoutCoordinates. Both mapped fail/pass/
review_required to a { resultComment, resultStatus } pair, differing
only in which RESULT_COMMENTS template they used and whether
addressDistance was included.

Fuses the previous evaluateAddressSimilarity helper and the two
switches into a single evaluateAddressSimilarityResult helper in
geolocation-and-address-precision.helpers.ts that takes the two
addresses plus a SimilarityCommentTemplates object (failed/passed/
reviewRequired callbacks) and returns the EvaluateResultOutput
directly. The switch logic now lives in exactly one place.

Each call site shrinks from a 38-line switch to a 20-line template
literal, and adding a new outcome kind in the future requires editing
only the helper. The SimilarityOutcome type and the per-call
similarityPercent destructuring are gone.
