+++ libs/methodologies/bold/rule-processors/mass-id/geolocation-and-address-precision/src/geolocation-and-address-precision.helpers.ts
@@ -1,4 +1,7 @@
-import { isNil, logger } from '@carrot-fndn/shared/helpers';
+import type { EvaluateResultOutput } from '@carrot-fndn/shared/rule/standard-data-processor';
+
+import { getEnableReviewRequired } from '@carrot-fndn/shared/env';
+import { isAddressMatch, isNil, logger } from '@carrot-fndn/shared/helpers';
 import { getEventAttributeValue } from '@carrot-fndn/shared/methodologies/bold/getters';
 import {
   getApprovedExceptions,
@@ -30,10 +33,49 @@ import type {
   GpsLongitudeApprovedException,
 } from './geolocation-and-address-precision.types';
 
+import { ADDRESS_SIMILARITY_THRESHOLD } from './geolocation-and-address-precision.constants';
 import {
   isGpsLatitudeApprovedException,
   isGpsLongitudeApprovedException,
 } from './geolocation-and-address-precision.validators';
+
+export interface SimilarityCommentTemplates {
+  failed: (similarityPercent: number) => string;
+  passed: (similarityPercent: number) => string;
+  reviewRequired: (similarityPercent: number) => string;
+}
+
+export const evaluateAddressSimilarityResult = (
+  eventAddress: DocumentAddress,
+  accreditedAddress: DocumentAddress,
+  templates: SimilarityCommentTemplates,
+): EvaluateResultOutput => {
+  const { isMatch, score } = isAddressMatch(
+    buildAddressComparisonString(eventAddress),
+    buildAddressComparisonString(accreditedAddress),
+    ADDRESS_SIMILARITY_THRESHOLD,
+  );
+  const similarityPercent = Math.floor(score * 100);
+
+  if (!isMatch || score < ADDRESS_SIMILARITY_THRESHOLD) {
+    return {
+      resultComment: templates.failed(similarityPercent),
+      resultStatus: 'FAILED',
+    };
+  }
+
+  if (getEnableReviewRequired()) {
+    return {
+      resultComment: templates.reviewRequired(similarityPercent),
+      resultStatus: 'REVIEW_REQUIRED',
+    };
+  }
+
+  return {
+    resultComment: templates.passed(similarityPercent),
+    resultStatus: 'PASSED',
+  };
+};
 
 export const buildAddressComparisonString = (
   address: DocumentAddress,
